Backend/requirements/requirement.txt
asgiref==3.8.1
astroid==3.3.5
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
dill==0.3.9
Django==5.1.2
django-cors-headers==4.6.0
django-filter==24.3
django-phonenumber-field==8.0.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
isort==5.13.2
Markdown==3.7
mccabe==0.7.0
networkx==3.4.2
oauthlib==3.2.2
packaging==24.1
phonenumbers==8.13.48
pillow==11.0.0
platformdirs==4.3.6
pycparser==2.22
PyJWT==2.9.0
pylint==3.3.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tomlkit==0.13.2
uritemplate==4.1.1
urllib3==2.2.3
